Move EthAddress to core and serde it without leading zeroes. (#97)
